requirements.txt
blinker
click
colorama
et-xmlfile
Flask
Flask-Cors
Flask-MySQLdb
google-cloud
google-cloud-bigquery
google-cloud-storage
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
mysqlclient
numpy
openpyxl
pandas
protobuf
PyMySQL
python-dateutil
python-dotenv
pytz
six
tzdata
Werkzeug